requirements.txt
youtube-transcript-api
# to fetch youtube data

python-docx
# to convert transcript into docs
